+++ back/src/server.ts
@@ -1,8 +1,7 @@
 import * as express from "express";
 import * as cors from "cors";
-import passswordResetRoutes from "./controller/passwordResetRoutes";
 import userController from "./controller/UserController";
-import { errorHandler } from "./middleware/errorHandler";
+import { errorHandler } from "./errors";
 
 const app = express();
 const PORT = 8080;
@@ -17,7 +16,6 @@ app.use(
 );
 
 // Rotas para redefinicao de senhaq
-app.use("/resetpassword", passswordResetRoutes);
 app.use("/user", userController);
 app.use(errorHandler);
